created a usecase dir in authoirzation, moved all usecases into the usecase dir
